+++ app/src/main/AndroidManifest.xml
@@ -11,20 +11,17 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme" >
-        <activity android:name=".SplashScreen" >
+        <activity android:name=".activity.SplashScreen" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".Google_Sign_in" />
-        <activity android:name=".bottom_bar">
-
+        <activity android:name=".activity.HomeActivity">
         </activity>
-        <activity android:name=".MainActivity" />
-        <activity android:name=".homeActivity" />
-        <activity android:name=".LoginScreen" >
+        <activity android:name=".activity.EditProfile"/>
+        <activity android:name=".activity.LoginScreen" >
         </activity>
 
         <meta-data
