--- app/src/main/java/com/dscnitp/touristshelperapp/Google_Sign_in.java
@@ -1,14 +1,0 @@
-package com.dscnitp.touristshelperapp;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class Google_Sign_in extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_google__sign_in);
-    }
-}
