--- app/src/main/java/com/dscnitp/touristshelperapp/MainActivity.java
@@ -1,30 +1,0 @@
-package com.dscnitp.touristshelperapp;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        Button button = (Button) findViewById(R.id.button_send);
-        button.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(view.getContext(),Google_Sign_in.class);
-                view.getContext().startActivity(intent);
-            }
-
-
-            });
-
-    }
-
-}
