+++ app/src/main/java/com/dscnitp/touristshelperapp/activity/EditProfile.java
@@ -1,0 +1,18 @@
+package com.dscnitp.touristshelperapp.activity;
+
+import android.os.Bundle;
+import android.os.PersistableBundle;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.dscnitp.touristshelperapp.R;
+
+public class EditProfile extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.edit_profile);
+    }
+}
