+++ app/src/main/java/com/dscnitp/touristshelperapp/fragment/ExploreFragment.java
@@ -1,0 +1,24 @@
+package com.dscnitp.touristshelperapp.fragment;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.dscnitp.touristshelperapp.R;
+
+
+public class ExploreFragment extends Fragment {
+
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_explore, container, false);
+    }
+}
