+++ app/src/main/java/com/dscnitp/touristshelperapp/fragment/LikesFragment.java
@@ -1,0 +1,22 @@
+package com.dscnitp.touristshelperapp.fragment;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.dscnitp.touristshelperapp.R;
+
+
+public class LikesFragment extends Fragment {
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_likes, container, false);
+    }
+}
