--- app/src/main/java/com/dscnitp/touristshelperapp/homeActivity.java
@@ -1,14 +1,0 @@
-package com.dscnitp.touristshelperapp;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class homeActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_home);
-    }
-}
