+++ app/src/main/res/layout/activity_bottom_bar.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".bottom_bar">
+    tools:context=".activity.HomeActivity">
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
