+++ app/src/main/res/layout/activity_login_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginScreen">
+    tools:context=".activity.LoginScreen">
 
     <Button
         android:id="@+id/google_btn"
