+++ app/src/main/res/layout/activity_login_with_otp.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginWithOtp">
+    tools:context=".activity.LoginWithOtp">
 
     <EditText
         android:id="@+id/phno"
