+++ app/src/main/res/layout/activity_profile.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:orientation="vertical"
     android:layout_height="match_parent"
-    tools:context=".Profile">
+    >
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -49,7 +49,7 @@
                         android:layout_width="146dp"
                         android:layout_height="151dp"
                         android:scaleType="centerCrop"
-                        android:src="@drawable/profile_img" />
+                        android:src="@drawable/ic_launcher_background" />
                 </androidx.cardview.widget.CardView>
                 <androidx.constraintlayout.widget.ConstraintLayout
                     android:layout_width="match_parent"
@@ -68,6 +68,18 @@
                         app:layout_constraintTop_toTopOf="parent"
                         app:layout_constraintVertical_bias="0.124"
                         app:srcCompat="@drawable/profile_setting" />
+
+                    <ImageView
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        app:layout_constraintTop_toTopOf="parent"
+                        app:layout_constraintEnd_toEndOf="parent"
+                        android:layout_marginEnd="20dp"
+                        android:padding="10dp"
+                        android:id="@+id/edit_profile"
+                        android:layout_marginTop="28dp"
+                        android:src="@drawable/ic_baseline_edit_24"
+                        />
 
                     <TextView
                         android:id="@+id/textView3"
@@ -92,7 +104,6 @@
                         android:layout_height="wrap_content"
                         android:layout_gravity="bottom"
                         android:text="@string/user_name"
-
                         android:textColor="#fff"
                         android:textSize="20sp"
                         android:textStyle="bold"
