+++ app/src/main/res/layout/fragment_maps.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MapsFragment">
+    tools:context=".fragment.MapsFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
